Fix race in reporting exit status over persistent connection
Summary:
When the server detects that .flowconfig has changed, then it naturally should exit with exit status FlowConfigChanged.

For the sake of LSP, it's not good for an LSP server to keep shutting down: VSCode and Nuclide will only restart it at most five times, and each time will be accompanied by a warning to the user.

So what we do is that for a small number of whitelisted reasons where it's normal and expected for the flow server to exit in the normal daily course of life, and where it's safe for the LSP adapter to simply restart the flow server, then it does precisely that. More precisely, the correct behavior of lspCommand.ml is unaffected by the contents of .flowconfig, but the choice of which flow binary to invoke is affected solely by the [version]. Therefore the rule is:

* If the server exits, and the exit was because of the whitelisted reason 'FlowConfig changed', and the [version] section of FlowConfig hasn't changed, then it's right and proper for lspCommand.ml to restart the flow server automatically.

To make this work, I'd changed the flow server so that immediately prior to exiting with ExitStatus.FlowConfigChanged, then it would send a message over the persistent connection saying "hey you, I'm about to fatally close the connection due to a FlowConfigChanged". (since the persistent-connection-client doesn't own the flow server process, it has no way of observing its exit stauts directly).

But there was a race in that. The job of the monitor is to relay messages from the server to the client, and also to listen for when the server shuts down and (if appropriate) terminate. Normally, the monitor was receiving the "hey I'm about to fatally close" message first, and forwarded this on to the client, and only picked up that the server had exited after this. But sometimes on a heavily-loaded machine, the monitor was picking up the server-exit first, and terminating abruptly, and so the "hey I'm about to fatally close" message never even got forwarded to the client.

This diff avoids the race by having the monitor be the party responsible for sending that message on to the client, not the server.

Reviewed By: gabelevi

Differential Revision: D7616745

fbshipit-source-id: ea8302ba41a2629d095d3061de7ba2f046944060