[hack] Expose Watchman socket fd so it can be selected on
Summary: If we start some lengthy rechecking and files change under our feet, we would like to cancel the unfinished part, process the changes, and then resume. Detection of this condition will happen using machinery added in {D6842414}. It uses `Unix.select` to monitor for interesting events, and in this diff I make Watchman socked fd available for this purpose.

Reviewed By: alexchow

Differential Revision: D7550942

fbshipit-source-id: b99b381c46920cc55bc71e3ffcaf2d7cb8e4920d