Pass multiple def-locs through find-refs
Summary:
We want the ability to search for references to any one of several different definition locations. This is a change which is internal to find-refs, it does not allow users to pass multiple locations to find-refs.

For the main motivating example, see the next diff, D7601765.

This change is also necessary for multi-hop find-refs.

Reviewed By: panagosg7

Differential Revision: D7588938

fbshipit-source-id: 51a61109947d26159607b3d126b37e0053329487