[Normalizer] Option to flag shadowed type parameters
Summary:
The normalizer keeps a stack of type parameters that are in scope. This stack
may contain the same name twice (but with different associated locations).
This is a case of shadowing. For certain uses of normalized types (e.g. suggest)
we do not wish to allow the generation of type parameters that are shadowed by
another definition. For example the inferred type for `z` in:
```
function outer<T>(y: T) {
  function inner<T>(x: T, z) {
    inner(x, y);
  }
}
```
is the _outer_ T. Adding the annotation `: T` for `z` would be incorrect. Before this
change this was allowed. This adds this optional behavior.

Reviewed By: samwgoldman

Differential Revision: D7658498

fbshipit-source-id: b05c808d96f2ac3551acb46c73bcd699d8f9832f