[PR] add covered_locs output to coverage
Summary:
Hello 👋

I've been working with the codecov team on the flow coverage report so I can upload my flow coverage report to http://codecov.io/ and use the service to analyze coverage over time.

One thing I was told was missing from the flow coverage report (in json format) was a `covered_locs` output.  There is `uncovered_locs` but no inverse for `covered_locs`, only the sum `covered_count` which makes it difficult for them to create the line-by-line report. Because the flow report only shows lines not covered Codecov never knows when lines are covered so it always results in 0% coverage. 🐼

This change simply adds `covered_locs` as an additional field to the json output.  I can't foresee that this could be a breaking change for anyone ingesting the json output. It does create slightly larger output files as all covered lines now need to be described as well.

I've only just given the latest coverage output to Codecov, when I hear back I'll update this if there are any additional issues but I wa
Closes https://github.com/facebook/flow/pull/3231

Reviewed By: samwgoldman, mrkev

Differential Revision: D4861141

Pulled By: avikchaudhuri

fbshipit-source-id: e0f8135539e43e46f52957602015de282f119bbd