[Normalizer] Add Module to Ty.t
Summary:
The module type did not have a counterpart in the Ty.t syntax (mostly because ModuleT
is a structure that encompasses a number of definitions, rather than a description of a set
of runtime values).

This commit introduces `Module name` in `Ty.t` to address the issue of having the normalizer
crash on `ModuleT`.

Reviewed By: mroch

Differential Revision: D7735113

fbshipit-source-id: 1c5d6525933898983f2186fb392654b46864f13b