union-find/disjoint-set data structure
Summary:
https://en.wikipedia.org/wiki/Disjoint-set_data_structure

This is needed for multi-hop find-refs in order to maintain sets of related types.

We already have an implementation of this data structure used to relate tvars, but it is not general at all and will be difficult to disentangle. I might try to move that over to here at some point but some initial efforts have shown that it won't be trivial.

Reviewed By: avikchaudhuri

Differential Revision: D7617677

fbshipit-source-id: 01d93e38771f08caa9fff09504823ab68232ea67