Optional chaining typing
Summary:
This adds type support for optional chaining. In `statement.ml`, whenever we reach a node representing a subscript operation (property access, method call, etc.), we recursively traverse the LHS until we reach the start of the property chain. As we do so, we build up a structure representing this chain. We can then walk this structure and emit constraints rather than recursively calling `expression` on the LHS. (We also have to update the type table.) This means we can maintain optional chaining state (specifically, are we short-circuiting?) in an iterative fashion rather than pass it around recursively via extra internal wrapper types.

Currently, we don't support method-like patterns in optional chains: if we encounter a Call whose callee is a Member where at least one of the Call/Member is OptionalCall/OptionalMember, we produce a Flow error.

Reviewed By: avikchaudhuri

Differential Revision: D7731783

fbshipit-source-id: b2b3cfb716b92b25e0fcdc1345ae10aa1a368daf