Factor out Flow-LSP conversions
Summary:
This moves flow-lsp conversions into a separate file.

It also fixes some "off-by-one" errors that I discovered by actually using it within Nuclide, where I'd misunderstood the 0-based or 1-based that Flow was using.

Note that lspCommand.ml will need to do some of them (for the things that are handled on the client, e.g. outline and for now errors). And commandHandler.ml will need to do the bulk of them.

Reviewed By: gabelevi

Differential Revision: D7644085

fbshipit-source-id: e6921acbf16d99b9ca1dce95144db7215ea24eb9