LSP supports errors
Summary:
This diff adds supports for `textDocument/publishDiagnostics`

My real goal is the server shouldn't send `Prot.Errors`, `Prot.StartRecheck` or `Prot.EndRecheck` over the persistent connection: it should send only an LSP notification to the client, to be handled in the normal way.

But that will take a heck of a lot of untangling that doesn't seem worth doing at the moment. So I settled for something much easier.

Reviewed By: gabelevi

Differential Revision: D7644150

fbshipit-source-id: 77c870f2a5b938430dcc3690252e9407dc9cb281