Strict Local
Summary:
`flow strict-local` is a mode like `flow strict`, except without the recursive `nonstrict-import` rule. It does not have a separate configuration, it is defined exactly as all the rules defined for strict mode, but the without the `nonstrict-import` rule.

For large existing codebases, this mode allows for engineers to enable stricter checks for a file before all the dependencies are `flow strict`. Once all the dependencies are `flow strict`, a `flow strict-local` file can always be updated to be `flow strict` by simply `s/flow strict/flow strict-local/`

A `flow strict` file cannot depend on a `flow strict-local` file, only on `flow strict` files.

Implementation note: We don't need to actually amend the lint rule list, as the `nonstrict-import` rule checks to see if the dependent is `flow strict` and only errors if it is and the dependency is not.

Reviewed By: avikchaudhuri

Differential Revision: D7552535

fbshipit-source-id: 82144cffd3b84212339063ae959f4f3c7b71fd25