simplify array spread handling
Summary:
When evaluating spreads inside an array literal, we would "finish" the array type with a different element type for each type that was being spread. This effectively created a "union of array types" instead of a single array type, which could be fanned into an exponential blowup, or worse, an infinite loop, by a suitably crafted cycle in the constraint graph.

Now we pin the element type early and thread it through so that we can always "finish" with that element type.

Fixes https://github.com/facebook/flow/issues/4070
Fixes https://github.com/facebook/flow/issues/4370

Reviewed By: samwgoldman

Differential Revision: D7416763

fbshipit-source-id: 7de51c426df341da9916c99a6e1eba0aec35a8fd