[hack] Don't send worker status to master for exit code 1
Summary:
When a flow server or flow check ends, we were seeing a bunch
of exceptions from the worker masters about trying to write to a broken
pipe. This diff fixes those errors.

It turns out, on unix systems, a broken pipe looks like a readable fd,
but when you read it you only get 0 bytes. Since we always read on the
worker slave, we need to fork before we realize that the pipe is broken.
The slave then exits with code 1. The worker master should not try to
send this status the master process, since the master is dead.

Reviewed By: alexchow

Differential Revision: D7727316

fbshipit-source-id: ea1942e84fedc0786ede7178967843ffe0f6f670