[hack] Expose `keytrace_to_string` in `Hh_json`, and make more readable
Summary:
Expose this in the case that we want to construct custom error messages for a certain keytrace.

This also improves the printing. Previously, it would wrap each key in square brackets, and the keys would be in reverse order, which is not an intuitive way of representing a keytrace through a JSON object.

Reviewed By: pittsw

Differential Revision: D7698999

fbshipit-source-id: e13ce89f1b41a822492843d059b8d6496b8f37ad