Avoid GetPropT on $call internal prop
Summary:
In a follow-up change, I want to move the $call property out of the props map,
and into an "internal slot" like proto_t. This is the only place in the code
where we do an operation (GetPropT) that only makes sense for properties.

I factored this change out from the larger change because it has an innocuous
change on error messages. In fact, I think the new error messages are nicer, as
they include more locations.

As noted inline, these rules are actually problematic. Because (1) `Class<B>`
is a subtype of `Class<A>` if `B` is a subclass of `A` and (2) call properties
and constructors are excluded from subclass compatibiltiy checks, it should be
illegal to call or construct a `Class<>` type. I hope to address this later.

Reviewed By: avikchaudhuri

Differential Revision: D7778585

fbshipit-source-id: 492b0dcd961a4599b64a8659efdb097ab306ec48