[hack] Use single int for file pos
Summary:
The `File_pos.t` type is currently a record of three Caml integers. Let's pack it into a single Caml integer to improve space usage for both `hh_single_compile` and `hh_server`.

We can do this and still provide very large ranges for the values:

* Use 30 bits for the "beginning of line offset", so max file size is 1G
* Use 24 bits for the line number, so max line number is 16M
* Use 9 bits for the column position, so max column position is 511

For the line and column, we can max out rather than wrap round or assert; it simply means that error messages degrade in these cases.

Differential Revision: D7774295

fbshipit-source-id: ca234d516853601b3fcc318b89ac3b58c7e540c7