Simplify GetStaticsT repositioning logic
Summary:
This logic feels overcomplicated to me.

1. We shouldn't need to wrap RStatics around the instance reason because the
reason on the statics member on an InstanceT was already created with the
appropriate reason.

2. There's simply too much repositioning and reason munging going on. We were
mucking with reasons before creating GetStaticsT, but it's simpler to just
reposition the statics type once.

Based on the test expected output change, I suspect this will improve positions
overall.

Reviewed By: panagosg7

Differential Revision: D7800036

fbshipit-source-id: d905eca6e8e023e197843572bd69a671220e5493