[Ty] Recursive types can be polymorphic
Summary:
This change addresses a shortcoming in normalizing polymorphic recursive types (formerly the
long comment in `type_app`). Consider the example:
```
type A<+T> = $ReadOnlyArray<A<T>> | T;
```
In normalizing the type for `A`, an `OpenT` is still "under resolution" as it appears as the
base of a `TypeAppT`.

Before, we handled this case by cheating and reading the name of the type in the `RType` description.

`RType`, however, is not a reliable source of information for the normalizer, so we would like to avoid
it wherever possible. At the same time, there is no reason why a recursive type variable cannot appear
at the base of a `Ty.Generic`. This change enables this.

This commit also introduces a wrapper `mk_mu` around the `Mu` constructor that performs
two checks before creating recursive types:
1. If the type is not really recursive (ie. the recursive variable does not appear in the body `t`) then
`t` is returned. This is possible after applying some simplifications/optimizations on an originally
seemingly recursive type.
2. The case `Mu (i, TypeAlias (n, ps, t)` is shortened to `TypeAlias (n, θ ps, θ t)`
where we've applied the substitution `θ = [i/n]`. This is a seemingly awkward fix, however, I am
not very concerned as `TypeAlias` is abusively treated as a type. Eventually, the goal of to include
it to a "module element" type rather than the `Ty.t` structure.

Reviewed By: samwgoldman

Differential Revision: D7798111

fbshipit-source-id: aa5546486a5d8596a8d162bc938841c3972dd4af