Ensure fields overwrite methods in interface definitions
Summary:
Classes (and declared classes) simultaneously define two shapes: own properties
and the prototype. Own properties are defined through fields while the
prototype is defined through methods.

For example, `class C { x: T, m(): U }` will create instances that have an
"own" shape of `{ x: T }` and a "proto" shape of `{ +m: () => U }`. Because
these two shapes are being defined at once, adding a method does not overwrite
a field and vice versa.

(Note that statics are an exception, since both static methods and static
fields are hosted on a single shape as own properties.)

However, for interfaces, we don't want the same behavior. We are moving toward
a model where interfaces are flat, so we want to be sure that fields overwrite
methods and methods overwrite fields.

(Note: multiple methods define an overridden method signature.)

Reviewed By: avikchaudhuri

Differential Revision: D7824827

fbshipit-source-id: 49fbb51922ffe92f1e20d5dc16f3f7dd6e29abcb