Add proto fields to the correct prop map
Summary:
There is currently a bit of confusion about own vs. proto properties on class
types. We are working toward a new object model where this distinction is very
important, so we need to be crystal clear on this.

I added support for parsing `proto` properties in declared classes in D7824119,
but the modifier was ignored until this diff. This diff ensures that the
properties are added to the `methods_tmap` field of the instance record, as
this prop map most closely models the proto properties of the class. I will
rename `fields_tmap` and `methods_tmap` in a follow-up to clarify the
distinction.

Reviewed By: avikchaudhuri

Differential Revision: D7825823

fbshipit-source-id: 26323d26d352a869fb2217966a840628fc9f90f8