Add parsing support for internal slot properties in object types
Summary:
It's useful to describe capabilities in object types, interfaces, and declared
classes which do not necessarily correspond to runtime properties.

Call properties are one example, which we support through a special syntax.
Unfortunately, this syntax is not unambiguous with methods when used in a
declare class -- `declare class C { static (): void }` can be reasonably parsed
as a static "callable" property or as an instance method named `static`.

Another example is "newable" objects, like constructors. It is useful to
describe the set of values which can be constructed with given arguments,
returning a value of some shape. We might support this with a syntax like
`{ [[new]](x: A, y: B): C }`.

Yet another example is `__proto__` property, which we currently support on
object types, but also potentially overlaps with a named property. We could
swap the syntax to the non-overlapping `[[proto]]` instead.

This diff simply adds parsing support for internal slot properties, and
provides no specific support. An internal slot property is currently always a
semantic error, realized during AST traversal. Follow-up diffs will add
support for specific properties.

Reviewed By: avikchaudhuri

Differential Revision: D7731906

fbshipit-source-id: 5051bc64c8c9317dff3196facff91c5a297539c3