Multi-hop find-refs
Summary:
Multi-hop find-refs determines when object types are related to each other via a subtyping relationship and links them. This way, a find-refs request for uses of a property on one type will include uses on its related types as well. For concrete examples look at the tests.

In theory this can be used to drive safe and precise global rename of properties in the presence of structural types. In practice it will require quite a lot of testing and probably bug fixing before it reaches that point.

Reviewed By: avikchaudhuri

Differential Revision: D7822298

fbshipit-source-id: 8ad54dbed518bf4be86ce613bb48fb0b78207a45