Parsing support for explicit type arguments in new/call
Summary:
This diff adds parsing support for explicit type arguments in new and call
expressions, e.g., `f<T>(x)`.

This syntax is overlaps with existing syntax in some cases. For example, the
program `f<T>(x)` already parses as nested binary expressions `((f<T)>x)`.

Flow will only attempt to parse this syntax as explicit type arguments if
`should_parse_types` is true -- i.e., if Flow is enabled for the file, either
through the `flow` pragma or the `--all` cmdline argument.

Due to the ambiguity, parsing this syntax uses backtracking via Try. When we
see `f<`, we try to parse the type argument list. If that fails, we rollback
and try parsing as a binary expression instead.

For optional chaining, the syntax `f?.<T>(x)` is parsed as an optional call
with type arguments.

Typing support for this syntax will be added in a follow-up. I have marked the
places where typing rules should consider type args with a TODO comment.

Reviewed By: gabelevi

Differential Revision: D7657631

fbshipit-source-id: 34390c43a10c30eeb18285f5282db502335f7419