substitution cache (part 1)
Summary:
Substitution is expensive in terms of space and time, especially when type definitions are large.

In this diff we avoid redundant substitution by caching the result of substituting a polymorphic
definition with a list of type args. The size of a cache key is small compared to the expanded
term. Particularly for files with large type definitions, this leads to a measurable perf win
(cutting redundant substitutions by multiple factors based on number of occurrences).

Reviewed By: samwgoldman

Differential Revision: D7804701

fbshipit-source-id: b450d373f8fa6c2d5a8d0c18eeaab39809646ae2