[hack] deriving show full_fidelity_source_text
Summary:
Here's an example of what we're doing to get deriving show available universally.

In order to minimize development pain, I suggest banning ppx extensions in mli files and
require users to write out the types and values they expect to see in full.

That way the magic is confined to source files and build errors are easier to detect.

Differential Revision: D7822111

fbshipit-source-id: fb34ea61fcecd5442f42fe099efe6a9057cdd838