Fix lsp/diagnostic flakey racey test
Summary:
This diff fixes a racey test.

The test has a step which waits for a "publishDiagnostics" message, and then checks which messages arrived in the step.

But Flow does incremental error reporting. There are different valid ways of doing that -- e.g. just send a single publishDiagnostics with the final results; or send a first publishDiagnostics that's empty followed by another one with final results; or a first publishDiagnostics that's identical to the final results followed by the final results.

Within a step waits for a "publishDiagnostics" message to arrive, if Flow happens to decide to send two publishDiagnostics messages, it's a race whether one or both will arrive in the step.

This diff makes the test robust against such races: we'll wait up to 9s to get the expected publishDiagnostic, then verify that this expected publishDiagnostic was sent at least once, and ignore any additional publishDiagnostics.

Reviewed By: nmote

Differential Revision: D7796512

fbshipit-source-id: 6b9b6002038dd423c459c8853fce9bb9a8909e9a