Misc lsp test fixes
Summary:
I've rewritten the `waitUntilXyz` functions in builder.js to be a little cleaner and to print more informative stuff in the test logs.

It used to always display the "timeout fired" message even if the wait condition had been satisfied before that. I removed it.

It used to run the `doneWithMsg` function body multiple times even when the wait condition had been satisfied before that. I stopped that.

There was a bug where `waitUntilServerStatus` was listening on the wrong emitter and so never fired until its timeout had been met. This resulted in lsp/connection test taking 20s longer than it should. I fixed this. (There were a few tests which relied upon this unecessary wait in order to reduce their flakiness; I fixed them up).

There was a bug where `waitUntilIDEMessage` wasn't examining all the messages that had so far been received in this step. This resulted in lsp/connection test taking 3s longer than it should. I fixed this.

The functions `waitUntilIDEMessage` and `waitUntilIDEMessageCount` used to throw an exception if the IDE wasn't operational at this stage. That's not in the spirit of the `waitUntil` functions, which merely return and rely upon a subsequent Verify function to check what happened. I changed them to return immediately in case the IDE is down, and report this in the log.

Reviewed By: gabelevi

Differential Revision: D7806971

fbshipit-source-id: 22f6b32b78e0218c03bed2f25d8857c419ecbfdd