LwtUtils.all
Summary:
This is basically the same as `Promise.all` and is consistent with Core's `Monad.S.all`.

It basically takes a list of promises and returns a promise for a list, which resolves to a list containing the elements that the individual promises resolved to. If any promise fails, the resulting promise fails immediately.

Reviewed By: gabelevi

Differential Revision: D7869782

fbshipit-source-id: 83d3feade14cdae7babccffbda7bc77c1b7ff6f6