Merge components in a sorted order
Summary:
Some of Flow's typechecking is a little racy. For example, errors in code with unsealed objects may or may not show up depending on the order that the flows are processed. Ideally, we'd find and fix all racy behavior. However, realistically it might be easier to make sure Flow always typechecks the code in the same order.

@@[5401580:samgoldman] suggested that one source of non-determinism is the order in which dependencies are merged. This diff attempts to remove that non-determinism by enforcing a strict order. This is done by replacing lists with maps & sets and iterating through those maps/sets instead.

Reviewed By: samwgoldman

Differential Revision: D7728354

fbshipit-source-id: b4676e8ebff5acd8b230dc3716e231bf242cfc6d