Allow reserved words as optional chain property names
Summary:
After parsing `?.` in `member_cover`, we peek one token ahead to anticipate whether we're dealing with a template literal, function type argument syntax, etc. I was using `is_identifier` to detect a dot-notation property access (`foo?.bar`), but this is too strict a check.

[[ https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#Reserved_word_usage | It is valid for reserved words to be used as property identifiers ]], but e.g. upon seeing `foo?.case`, the `is_identifier` check would fail due to `case` having the token `T_CASE` instead of `T_IDENTIFIER`.

The easy solution is to restructure the lookahead to have `foo?.bar` be the default case rather than try to lookahead for it.

Reviewed By: gabelevi

Differential Revision: D7859446

fbshipit-source-id: 553692e920ae6e7b1bc17309dc60197daf3e7f92