Rework Func_params module
Summary:
The purpose of Func_params (and Func_sig, Class_sig, and Default) is to hold on
to structures that needs to be substituted before we can emit any constraints,
to avoid emitting constraints involving BoundT.

These modules each contain a small data structure and helpers that operate on
that data, but also contain code to create those data from ASTs.

In a follow-up, I want to create interfaces from type_annotation.ml, which
necessitates the use of these modules. However, the conversion code creates a
recursive dependency between type_annotation.ml and the func_params.ml.

This diff is the first in a series of changes to break the dependencies on
type_annotation.ml from these modules. Instead, these modules should only
include the data structure and helper functions. The rest of the code will be
moved to statement.ml or and type_annotation.ml.

This diff doesn't move anything. This diff reworks the (abstract) Func_params.t
data type and adds public methods which can be shared between the two cases we
care about: building from type and value ASTs.

Reviewed By: panagosg7

Differential Revision: D7886723

fbshipit-source-id: f1e0cb7ff50a4a99c6cd66467cf7217acecaca82