carry dependency graph in env
Summary:
We carry around the set of parsed files, so why not carry around the dependency graph over that set
as well?

In particular, this means that the code path for typecheck_contents, which all of our commands hit,
gets optimized: it can look up the dependency graph instead of recomputing it.

Reviewed By: gabelevi

Differential Revision: D7962614

fbshipit-source-id: 06127f032e624b024c56755fa8d2ba71a768611d