speed up reversing dependency graph
Summary:
Reversing a graph can be expensive if not done carefully. In particular, it is a bad idea to use a
map of sets to accumulate the reverse graph, because we might end up doing lookups and additions
that are proportional to the number of edges (quadratic in the number of nodes), scaled by a log
factor. Roughly speaking.

Instead, using a hash table takes time linear in the number of edges. In particular, OCaml's hash
table exposes an API to `add` and `find_all` bindings mapping to the same key, which can be
exploited to accumulate and retrieve reverse edges easily.

On large code bases, the saved log factor can be noticeable. Also uses less memory, so less GC.

Reviewed By: gabelevi

Differential Revision: D7963041

fbshipit-source-id: 336f04752925ed7e0cdad2c1fbc14299f74147fd