incremental compute of dependency graph
Summary:
Rather than recompute the entire dependency graph on every recheck, update the old dependency graph,
deleting some entries and recomputing other entries based on a smaller recomputation of the
dependency graph for new and changed files. This scales as the size of the update rather than the
size of the code base.

Reviewed By: gabelevi

Differential Revision: D7963215

fbshipit-source-id: 547b5c085d5376d675a61bbefb46150432576e7d