[hack] Priority pipe between server and monitor
Summary:
Some (most) commands don't need full recheck, but dumping them in same pipe with those that do makes it hard to serve them as quickly as they could (because we have no control what kind of client arrives, and after it does it's too late - we are already using resources to handle it, running out of fds, and other horrible stuff).

In this diff I introduce additional pipe that can be used to queue different kind of queries in different places and avoid this problem.

Reviewed By: alexchow

Differential Revision: D7923566

fbshipit-source-id: e2760e14c17a939dd5f18cb19c95495e16c8077b