Update compiler-babel.md
I've seen a few instances of people using the `flow` as well as `react` presets unnecessarily. Since many users will already be using the React preset is it worth pointing this out here?

Or has a conscious effort been made to not mention React to counter the preconception that Flow can only be used with React?