Make T_INTERFACE and T_EXTENDS type tokens
Summary:
In a follow-up, I add parsing support for interface types, which are anonymous
interfaces that can appear anywhere a type can.

This means we will start tokenizing `interface` and `extends` in the TYPE lex
mode as the T_INTERFACE and T_EXTENDS tokens, respectively.

This is a slight breaking change, because we used to successfully parse the
statement `type interface = ...`. Note that `type extends = ...` is already an
error, but this diff also makes it an error to refer to a type named `extends`.

Reviewed By: gabelevi

Differential Revision: D7885142

fbshipit-source-id: 94178c2e12318c8c2a7669a5fea0fe3cba2eedc6