Typing support for inline interfaces
Summary:
An inline interface is created in nearly the exact same way as a named
interface declaration, creating a InstanceT type with the structural flag set
to true. Thus, all the subtyping rules are standard and expected.

There are only a few differences:

(1) Inline interfaces don't have an implicit `name` property. Soon, named
interfaces will not either. This was an accidentally inherited behavior from
code sharing with declared classes.

(2) Inline interfaces are checked for polarity. This behavior comes "for free"
since all type aliases are checked for polarity.

Reviewed By: panagosg7

Differential Revision: D7887560

fbshipit-source-id: e6a315e7dc56ac7bf0da02a19575d51ead791800