Allow internal slot properties to be optional
Summary:
This will come in handy with `[[call]]`, which can be used to define a type
that is not callable. `{ [[call]]?: empty }` is the type of all non-callable
objects. This is useful for React's setState argument, which is either a
function or a shape.

Currently we have some ugly hacks to ensure that function types are
incompatible with $Shape and dictionaries. This will give us a more principled
way to express these APIs.

I will follow up with Babylon to add support for this there as well.

Reviewed By: gabelevi

Differential Revision: D8041629

fbshipit-source-id: b058527b6577655a4dfcbc631757fe576797d07b