O(1) stack membership lookup
Summary:
Checking stack membership was linear in the size of the stack. Since the stack can grow as the size
of cycles, this operation can become super expensive on large graphs with lots of cycles.

We can pay for this cost with a bit of extra memory and an O(1) check. This mimics what happens in
the original Tarjan algorithm.

Reviewed By: gabelevi

Differential Revision: D8000896

fbshipit-source-id: 2f1bd5be753dcef08a2b05f8b2cabdb2ab13f018