Use precise location of imported name for member access hook and type table
Summary:
`import {bar as baz} from ...`

Before this pointed to the entire `bar as baz` range. Now it's confined to `bar`
-- which is the actual location of the identifier acting as the member access.

This leads to some minor changes in get-def behavior. Now, get-def on the `as`
or `baz` in the above example points to the beginning of the module (same as
anywhere within the `import` statement). get-def on the `bar` still jumps to the
actual location where `bar` is exported. Jumping through the import still works
as expected.

Reviewed By: samwgoldman

Differential Revision: D8049240

fbshipit-source-id: 8c2bf29da1f4851e585df1659cee01550466038c