reuse topsort instead of recalculating it
Summary:
Topological sort of a dependency graph, as currently implemented, is not cheap for large graphs
(it's supposed to be linear in the number of edges, but I suspect it's not because of the use of
maps and sets).

Currently we do it twice on the same code path. This diff does it once and then uses the result to
derive the other result. This is based on a strong invariant on the closure properties of the root
set and the dependency graph on which the topological sort is computed. I'll make this invariant
more obvious in a follow-up diff.

Reviewed By: gabelevi

Differential Revision: D7981130

fbshipit-source-id: 4ac61c43c2bbf5cd99a4dfdb32272d69319ab9c6