reduce make_merge_input to prep_merge
Summary:
MakeMergeInput seems like it's doing something substantial, but it's not. It turns out that calling
the callback `make_merge_input` is not necessary to determine its output. As such, moving the
computational part out of the callback makes the data flow more explicit. In particular, we can more
easily observe that `to_merge = infer_input + all_dependent_files`, the key set of `recheck_map` is
`to_merge`, and the values of `recheck_map` are determined by membership in `roots = infer_input +
direct_dependent_files`.

Reviewed By: gabelevi

Differential Revision: D7995488

fbshipit-source-id: 0a91b4de40e4e4533002eb8277cdc0c5aaa10359