don't rely on removing explored edges in topsort
Summary:
Removing explored edges was a way to ensure that edges are not explored twice. But this operation is
not only O(log N), which can be expensive for large graphs, but also creates intermediate maps which
causes memory pressure.

Since we already track which nodes have been visited, we can do away with this operation, replacing
it with a O(1) membership check.

Reviewed By: gabelevi

Differential Revision: D8000850

fbshipit-source-id: ae6cd7d757dff7829ad1bb60b2cfbb109ae0aaab