Associate the function name in `export default async function...` with the export
Summary:
Unlike `export default function...`, `export default async function` is treated
as the export of an arbitrary expression which just happens to evaluate to a
function. This means that no top-level name is introduced.

However, we still want to allow the function name to be a target for
`find-refs`, so in the special-case of a default export of a function
expression, let's mark the name as a local binding. Of course, this name is not
available in the top-level scope but it can be called recursively from within
the function so there still may be references.

Also added a test for another default exports of an arbitrary expression.

Reviewed By: panagosg7

Differential Revision: D8123222

fbshipit-source-id: 77f2694cb863b2c517f7c9f5307b508f04060f6b